refs #15207 [serveur] Access the user from WebSocket messages
Since the user is stored in the session (which uses Mongo), we need to get access to the session ID. The only way I found to solve this issue is by accessing directly the cookies sent by the client and looking for the connect.sid. It's ugly but it appears to be working.