refs #16881 [lourd]  Pixel Mode: Huge Rename + Selector Mode +  Eraser Cursor
Don't let the commit size scaring you, it's simply cause by file rename.

I forgot the commit individually two small changes here. So they mingle with this commit The first is the eraser cursor on PixelEditor. The second is the OpenEditorWindow that choose which Editor open.

Knowing this commit would be that huge, I would have made him in another PR.  I'll remember it next time.